+++ src/main/java/kitchenpos/deliveryorders/application/DeliveryOrderService.java
@@ -1,0 +1,143 @@
+package kitchenpos.deliveryorders.application;
+
+import kitchenpos.deliveryorders.domain.*;
+import kitchenpos.menus.tobe.domain.menu.Menu;
+import kitchenpos.menus.tobe.domain.menu.MenuRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+import java.util.*;
+import java.util.stream.Collectors;
+
+@Service
+public class DeliveryOrderService {
+    private final DeliveryOrderRepository orderRepository;
+    private final MenuRepository menuRepository;
+    private final KitchenridersClient kitchenridersClient;
+
+    public DeliveryOrderService(
+            final DeliveryOrderRepository orderRepository,
+            final MenuRepository menuRepository,
+            final KitchenridersClient kitchenridersClient
+    ) {
+        this.orderRepository = orderRepository;
+        this.menuRepository = menuRepository;
+        this.kitchenridersClient = kitchenridersClient;
+    }
+
+    @Transactional
+    public DeliveryOrder create(final DeliveryOrder request) {
+        final List<DeliveryOrderLineItem> orderLineItemRequests = request.getOrderLineItems();
+        if (Objects.isNull(orderLineItemRequests) || orderLineItemRequests.isEmpty()) {
+            throw new IllegalArgumentException();
+        }
+        final List<Menu> menus = menuRepository.findAllByIdIn(
+                orderLineItemRequests.stream()
+                        .map(DeliveryOrderLineItem::getMenuId)
+                        .collect(Collectors.toList())
+        );
+        if (menus.size() != orderLineItemRequests.size()) {
+            throw new IllegalArgumentException();
+        }
+        final List<DeliveryOrderLineItem> orderLineItems = new ArrayList<>();
+        for (final DeliveryOrderLineItem orderLineItemRequest : orderLineItemRequests) {
+            final long quantity = orderLineItemRequest.getQuantity();
+            if (quantity < 0) {
+                throw new IllegalArgumentException();
+            }
+            final Menu menu = menuRepository.findById(orderLineItemRequest.getMenuId())
+                    .orElseThrow(NoSuchElementException::new);
+            if (!menu.isDisplayed()) {
+                throw new IllegalStateException();
+            }
+            if (menu.getPrice().getValue().compareTo(orderLineItemRequest.getPrice()) != 0) {
+                throw new IllegalArgumentException();
+            }
+            final DeliveryOrderLineItem orderLineItem = new DeliveryOrderLineItem();
+            orderLineItem.setMenu(menu);
+            orderLineItem.setQuantity(quantity);
+            orderLineItems.add(orderLineItem);
+        }
+        DeliveryOrder order = new DeliveryOrder();
+        order.setId(UUID.randomUUID());
+        order.setStatus(OrderStatus.WAITING);
+        order.setOrderDateTime(LocalDateTime.now());
+        order.setOrderLineItems(orderLineItems);
+        final String deliveryAddress = request.getDeliveryAddress();
+        if (Objects.isNull(deliveryAddress) || deliveryAddress.isEmpty()) {
+            throw new IllegalArgumentException();
+        }
+        order.setDeliveryAddress(deliveryAddress);
+        return orderRepository.save(order);
+    }
+
+    @Transactional
+    public DeliveryOrder accept(final UUID orderId) {
+        final DeliveryOrder order = orderRepository.findById(orderId)
+                .orElseThrow(NoSuchElementException::new);
+        if (order.getStatus() != OrderStatus.WAITING) {
+            throw new IllegalStateException();
+        }
+        BigDecimal sum = BigDecimal.ZERO;
+        for (final DeliveryOrderLineItem orderLineItem : order.getOrderLineItems()) {
+            sum = orderLineItem.getMenu()
+                    .getPrice().getValue()
+                    .multiply(BigDecimal.valueOf(orderLineItem.getQuantity()));
+        }
+        kitchenridersClient.requestDelivery(orderId, sum, order.getDeliveryAddress());
+        order.setStatus(OrderStatus.ACCEPTED);
+        return order;
+    }
+
+    @Transactional
+    public DeliveryOrder serve(final UUID orderId) {
+        final DeliveryOrder order = orderRepository.findById(orderId)
+                .orElseThrow(NoSuchElementException::new);
+        if (order.getStatus() != OrderStatus.ACCEPTED) {
+            throw new IllegalStateException();
+        }
+        order.setStatus(OrderStatus.SERVED);
+        return order;
+    }
+
+    @Transactional
+    public DeliveryOrder startDelivery(final UUID orderId) {
+        final DeliveryOrder order = orderRepository.findById(orderId)
+                .orElseThrow(NoSuchElementException::new);
+        if (order.getStatus() != OrderStatus.SERVED) {
+            throw new IllegalStateException();
+        }
+        order.setStatus(OrderStatus.DELIVERING);
+        return order;
+    }
+
+    @Transactional
+    public DeliveryOrder completeDelivery(final UUID orderId) {
+        final DeliveryOrder order = orderRepository.findById(orderId)
+                .orElseThrow(NoSuchElementException::new);
+        if (order.getStatus() != OrderStatus.DELIVERING) {
+            throw new IllegalStateException();
+        }
+        order.setStatus(OrderStatus.DELIVERED);
+        return order;
+    }
+
+    @Transactional
+    public DeliveryOrder complete(final UUID orderId) {
+        final DeliveryOrder order = orderRepository.findById(orderId)
+                .orElseThrow(NoSuchElementException::new);
+        final OrderStatus status = order.getStatus();
+        if (status != OrderStatus.DELIVERED) {
+            throw new IllegalStateException();
+        }
+        order.setStatus(OrderStatus.COMPLETED);
+        return order;
+    }
+
+    @Transactional(readOnly = true)
+    public List<DeliveryOrder> findAll() {
+        return orderRepository.findAll();
+    }
+}
