+++ src/main/java/kitchenpos/deliveryorders/domain/KitchenridersClient.java
@@ -1,0 +1,8 @@
+package kitchenpos.deliveryorders.domain;
+
+import java.math.BigDecimal;
+import java.util.UUID;
+
+public interface KitchenridersClient {
+    void requestDelivery(UUID orderId, BigDecimal amount, String deliveryAddress);
+}
