+++ src/main/java/kitchenpos/deliveryorders/domain/OrderStatus.java
@@ -1,0 +1,5 @@
+package kitchenpos.deliveryorders.domain;
+
+public enum OrderStatus {
+    WAITING, ACCEPTED, SERVED, DELIVERING, DELIVERED, COMPLETED
+}
