+++ src/main/java/kitchenpos/deliveryorders/infrastructure/JpaDeliveryOrderRepository.java
@@ -1,0 +1,10 @@
+package kitchenpos.deliveryorders.infrastructure;
+
+import kitchenpos.deliveryorders.domain.DeliveryOrder;
+import kitchenpos.deliveryorders.domain.DeliveryOrderRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface JpaDeliveryOrderRepository extends DeliveryOrderRepository, JpaRepository<DeliveryOrder, UUID> {
+}
