+++ src/main/java/kitchenpos/eatinorders/application/orderTables/OrderTableService.java
@@ -1,0 +1,78 @@
+package kitchenpos.eatinorders.application.orderTables;
+
+import kitchenpos.eatinorders.domain.orders.EatInOrderRepository;
+import kitchenpos.eatinorders.domain.orders.OrderStatus;
+import kitchenpos.eatinorders.domain.ordertables.OrderTable;
+import kitchenpos.eatinorders.domain.ordertables.OrderTableRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Objects;
+import java.util.UUID;
+
+@Service
+public class OrderTableService {
+    private final OrderTableRepository orderTableRepository;
+    private final EatInOrderRepository eatInOrderRepository;
+
+    public OrderTableService(final OrderTableRepository orderTableRepository, final EatInOrderRepository eatInOrderRepository) {
+        this.orderTableRepository = orderTableRepository;
+        this.eatInOrderRepository = eatInOrderRepository;
+    }
+
+    @Transactional
+    public OrderTable create(final OrderTable request) {
+        final String name = request.getName();
+        if (Objects.isNull(name) || name.isEmpty()) {
+            throw new IllegalArgumentException();
+        }
+        final OrderTable orderTable = new OrderTable();
+        orderTable.setId(UUID.randomUUID());
+        orderTable.setName(name);
+        orderTable.setNumberOfGuests(0);
+        orderTable.setOccupied(false);
+        return orderTableRepository.save(orderTable);
+    }
+
+    @Transactional
+    public OrderTable sit(final UUID orderTableId) {
+        final OrderTable orderTable = orderTableRepository.findById(orderTableId)
+                .orElseThrow(NoSuchElementException::new);
+        orderTable.setOccupied(true);
+        return orderTable;
+    }
+
+    @Transactional
+    public OrderTable clear(final UUID orderTableId) {
+        final OrderTable orderTable = orderTableRepository.findById(orderTableId)
+                .orElseThrow(NoSuchElementException::new);
+        if (eatInOrderRepository.existsByOrderTableAndStatusNot(orderTable, OrderStatus.COMPLETED)) {
+            throw new IllegalStateException();
+        }
+        orderTable.setNumberOfGuests(0);
+        orderTable.setOccupied(false);
+        return orderTable;
+    }
+
+    @Transactional
+    public OrderTable changeNumberOfGuests(final UUID orderTableId, final OrderTable request) {
+        final int numberOfGuests = request.getNumberOfGuests();
+        if (numberOfGuests < 0) {
+            throw new IllegalArgumentException();
+        }
+        final OrderTable orderTable = orderTableRepository.findById(orderTableId)
+                .orElseThrow(NoSuchElementException::new);
+        if (!orderTable.isOccupied()) {
+            throw new IllegalStateException();
+        }
+        orderTable.setNumberOfGuests(numberOfGuests);
+        return orderTable;
+    }
+
+    @Transactional(readOnly = true)
+    public List<OrderTable> findAll() {
+        return orderTableRepository.findAll();
+    }
+}
