+++ src/main/java/kitchenpos/eatinorders/application/orders/EatInOrderService.java
@@ -1,0 +1,120 @@
+package kitchenpos.eatinorders.application.orders;
+
+import kitchenpos.eatinorders.domain.orders.EatInOrder;
+import kitchenpos.eatinorders.domain.orders.EatInOrderLineItem;
+import kitchenpos.eatinorders.domain.orders.EatInOrderRepository;
+import kitchenpos.eatinorders.domain.orders.OrderStatus;
+import kitchenpos.eatinorders.domain.ordertables.OrderTable;
+import kitchenpos.eatinorders.domain.ordertables.OrderTableRepository;
+import kitchenpos.menus.tobe.domain.menu.Menu;
+import kitchenpos.menus.tobe.domain.menu.MenuRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.time.LocalDateTime;
+import java.util.*;
+import java.util.stream.Collectors;
+
+@Service
+public class EatInOrderService {
+    private final EatInOrderRepository eatInOrderRepository;
+    private final MenuRepository menuRepository;
+    private final OrderTableRepository orderTableRepository;
+
+    public EatInOrderService(
+            final EatInOrderRepository eatInOrderRepository,
+            final MenuRepository menuRepository,
+            final OrderTableRepository orderTableRepository
+    ) {
+        this.eatInOrderRepository = eatInOrderRepository;
+        this.menuRepository = menuRepository;
+        this.orderTableRepository = orderTableRepository;
+    }
+
+    @Transactional
+    public EatInOrder create(final EatInOrder request) {
+        final List<EatInOrderLineItem> orderLineItemRequests = request.getOrderLineItems();
+        if (Objects.isNull(orderLineItemRequests) || orderLineItemRequests.isEmpty()) {
+            throw new IllegalArgumentException();
+        }
+        final List<Menu> menus = menuRepository.findAllByIdIn(
+                orderLineItemRequests.stream()
+                        .map(EatInOrderLineItem::getMenuId)
+                        .collect(Collectors.toList())
+        );
+        if (menus.size() != orderLineItemRequests.size()) {
+            throw new IllegalArgumentException();
+        }
+        final List<EatInOrderLineItem> orderLineItems = new ArrayList<>();
+        for (final EatInOrderLineItem orderLineItemRequest : orderLineItemRequests) {
+            final Menu menu = menuRepository.findById(orderLineItemRequest.getMenuId())
+                    .orElseThrow(NoSuchElementException::new);
+            if (!menu.isDisplayed()) {
+                throw new IllegalStateException();
+            }
+            if (menu.getPrice().getValue().compareTo(orderLineItemRequest.getPrice()) != 0) {
+                throw new IllegalArgumentException();
+            }
+            final EatInOrderLineItem orderLineItem = new EatInOrderLineItem();
+            orderLineItem.setMenu(menu);
+            orderLineItem.setQuantity(orderLineItemRequest.getQuantity());
+            orderLineItems.add(orderLineItem);
+        }
+        EatInOrder eatInOrder = new EatInOrder();
+        eatInOrder.setId(UUID.randomUUID());
+        eatInOrder.setStatus(OrderStatus.WAITING);
+        eatInOrder.setOrderDateTime(LocalDateTime.now());
+        eatInOrder.setOrderLineItems(orderLineItems);
+        final OrderTable orderTable = orderTableRepository.findById(request.getOrderTableId())
+                .orElseThrow(NoSuchElementException::new);
+        if (!orderTable.isOccupied()) {
+            throw new IllegalStateException();
+        }
+        eatInOrder.setOrderTable(orderTable);
+        return eatInOrderRepository.save(eatInOrder);
+    }
+
+    @Transactional
+    public EatInOrder accept(final UUID orderId) {
+        final EatInOrder eatInOrder = eatInOrderRepository.findById(orderId)
+                .orElseThrow(NoSuchElementException::new);
+        if (eatInOrder.getStatus() != OrderStatus.WAITING) {
+            throw new IllegalStateException();
+        }
+        eatInOrder.setStatus(OrderStatus.ACCEPTED);
+        return eatInOrder;
+    }
+
+    @Transactional
+    public EatInOrder serve(final UUID orderId) {
+        final EatInOrder eatInOrder = eatInOrderRepository.findById(orderId)
+                .orElseThrow(NoSuchElementException::new);
+        if (eatInOrder.getStatus() != OrderStatus.ACCEPTED) {
+            throw new IllegalStateException();
+        }
+        eatInOrder.setStatus(OrderStatus.SERVED);
+        return eatInOrder;
+    }
+
+    @Transactional
+    public EatInOrder complete(final UUID orderId) {
+        final EatInOrder eatInOrder = eatInOrderRepository.findById(orderId)
+                .orElseThrow(NoSuchElementException::new);
+        final OrderStatus status = eatInOrder.getStatus();
+        if (status != OrderStatus.SERVED) {
+            throw new IllegalStateException();
+        }
+        eatInOrder.setStatus(OrderStatus.COMPLETED);
+        final OrderTable orderTable = eatInOrder.getOrderTable();
+        if (!eatInOrderRepository.existsByOrderTableAndStatusNot(orderTable, OrderStatus.COMPLETED)) {
+            orderTable.setNumberOfGuests(0);
+            orderTable.setOccupied(false);
+        }
+        return eatInOrder;
+    }
+
+    @Transactional(readOnly = true)
+    public List<EatInOrder> findAll() {
+        return eatInOrderRepository.findAll();
+    }
+}
