+++ src/main/java/kitchenpos/eatinorders/domain/orders/EatInOrder.java
@@ -1,0 +1,94 @@
+package kitchenpos.eatinorders.domain.orders;
+
+import kitchenpos.eatinorders.domain.ordertables.OrderTable;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.UUID;
+
+@Table(name = "eat_in_orders")
+@Entity
+public class EatInOrder {
+    @Column(name = "id", columnDefinition = "binary(16)")
+    @Id
+    private UUID id;
+
+    @Column(name = "status", nullable = false)
+    @Enumerated(EnumType.STRING)
+    private OrderStatus status;
+
+    @Column(name = "order_date_time", nullable = false)
+    private LocalDateTime orderDateTime;
+
+    @OneToMany(cascade = {CascadeType.PERSIST, CascadeType.MERGE})
+    @JoinColumn(
+            name = "order_id",
+            nullable = false,
+            columnDefinition = "binary(16)",
+            foreignKey = @ForeignKey(name = "fk_eat_in_order_line_item_to_eat_in_orders")
+    )
+    private List<EatInOrderLineItem> orderLineItems;
+
+    @ManyToOne
+    @JoinColumn(
+            name = "order_table_id",
+            columnDefinition = "binary(16)",
+            foreignKey = @ForeignKey(name = "fk_orders_to_order_table")
+    )
+    private OrderTable orderTable;
+
+    @Transient
+    private UUID orderTableId;
+
+    public EatInOrder() {
+    }
+
+    public UUID getId() {
+        return id;
+    }
+
+    public void setId(final UUID id) {
+        this.id = id;
+    }
+
+    public OrderStatus getStatus() {
+        return status;
+    }
+
+    public void setStatus(final OrderStatus status) {
+        this.status = status;
+    }
+
+    public LocalDateTime getOrderDateTime() {
+        return orderDateTime;
+    }
+
+    public void setOrderDateTime(final LocalDateTime orderDateTime) {
+        this.orderDateTime = orderDateTime;
+    }
+
+    public List<EatInOrderLineItem> getOrderLineItems() {
+        return orderLineItems;
+    }
+
+    public void setOrderLineItems(List<EatInOrderLineItem> orderLineItems) {
+        this.orderLineItems = orderLineItems;
+    }
+
+    public OrderTable getOrderTable() {
+        return orderTable;
+    }
+
+    public void setOrderTable(final OrderTable orderTable) {
+        this.orderTable = orderTable;
+    }
+
+    public UUID getOrderTableId() {
+        return orderTableId;
+    }
+
+    public void setOrderTableId(final UUID orderTableId) {
+        this.orderTableId = orderTableId;
+    }
+}
