+++ src/main/java/kitchenpos/eatinorders/domain/orders/OrderStatus.java
@@ -1,0 +1,5 @@
+package kitchenpos.eatinorders.domain.orders;
+
+public enum OrderStatus {
+    WAITING, ACCEPTED, SERVED, COMPLETED
+}
