+++ src/main/java/kitchenpos/eatinorders/domain/ordertables/OrderTableRepository.java
@@ -1,0 +1,14 @@
+package kitchenpos.eatinorders.domain.ordertables;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
+public interface OrderTableRepository {
+    OrderTable save(OrderTable orderTable);
+
+    Optional<OrderTable> findById(UUID id);
+
+    List<OrderTable> findAll();
+}
+
