+++ src/main/java/kitchenpos/eatinorders/infrastructure/JpaEatInOrderRepository.java
@@ -1,0 +1,10 @@
+package kitchenpos.eatinorders.infrastructure;
+
+import kitchenpos.eatinorders.domain.orders.EatInOrder;
+import kitchenpos.eatinorders.domain.orders.EatInOrderRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface JpaEatInOrderRepository extends EatInOrderRepository, JpaRepository<EatInOrder, UUID> {
+}
