+++ src/main/java/kitchenpos/eatinorders/infrastructure/JpaOrderTableRepository.java
@@ -1,0 +1,10 @@
+package kitchenpos.eatinorders.infrastructure;
+
+import kitchenpos.eatinorders.domain.ordertables.OrderTable;
+import kitchenpos.eatinorders.domain.ordertables.OrderTableRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface JpaOrderTableRepository extends OrderTableRepository, JpaRepository<OrderTable, UUID> {
+}
