+++ src/main/java/kitchenpos/eatinorders/ui/EatInOrderRestController.java
@@ -1,0 +1,47 @@
+package kitchenpos.eatinorders.ui;
+
+import kitchenpos.eatinorders.application.orders.EatInOrderService;
+import kitchenpos.eatinorders.domain.orders.EatInOrder;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+import java.util.List;
+import java.util.UUID;
+
+@RequestMapping("/api/eat-in-orders")
+@RestController
+public class EatInOrderRestController {
+    private final EatInOrderService eatInOrderService;
+
+    public EatInOrderRestController(EatInOrderService eatInOrderService) {
+        this.eatInOrderService = eatInOrderService;
+    }
+
+    @PostMapping
+    public ResponseEntity<EatInOrder> create(@RequestBody final EatInOrder request) {
+        final EatInOrder response = eatInOrderService.create(request);
+        return ResponseEntity.created(URI.create("/api/orders/" + response.getId()))
+                .body(response);
+    }
+
+    @PutMapping("/{orderId}/accept")
+    public ResponseEntity<EatInOrder> accept(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(eatInOrderService.accept(orderId));
+    }
+
+    @PutMapping("/{orderId}/serve")
+    public ResponseEntity<EatInOrder> serve(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(eatInOrderService.serve(orderId));
+    }
+
+    @PutMapping("/{orderId}/complete")
+    public ResponseEntity<EatInOrder> complete(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(eatInOrderService.complete(orderId));
+    }
+
+    @GetMapping
+    public ResponseEntity<List<EatInOrder>> findAll() {
+        return ResponseEntity.ok(eatInOrderService.findAll());
+    }
+}
