+++ src/main/java/kitchenpos/takeoutorders/domain/OrderStatus.java
@@ -1,0 +1,5 @@
+package kitchenpos.takeoutorders.domain;
+
+public enum OrderStatus {
+    WAITING, ACCEPTED, SERVED, COMPLETED
+}
