+++ src/main/java/kitchenpos/takeoutorders/infrastructure/JpaTakeoutOrderRepository.java
@@ -1,0 +1,10 @@
+package kitchenpos.takeoutorders.infrastructure;
+
+import kitchenpos.takeoutorders.domain.TakeoutOrder;
+import kitchenpos.takeoutorders.domain.TakeoutOrderRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface JpaTakeoutOrderRepository extends TakeoutOrderRepository, JpaRepository<TakeoutOrder, UUID> {
+}
