+++ src/test/java/kitchenpos/Fixtures.java
@@ -1,6 +1,6 @@
 package kitchenpos;
 
-import kitchenpos.eatinorders.domain.*;
+import kitchenpos.eatinorders.domain.ordertables.OrderTable;
 import kitchenpos.menus.application.FakeMenuNameProfanities;
 import kitchenpos.menus.tobe.domain.menu.*;
 import kitchenpos.menus.tobe.domain.menugroup.MenuGroup;
@@ -11,9 +11,7 @@ import kitchenpos.products.application.InMemoryProductRepository;
 import kitchenpos.products.tobe.domain.*;
 
 import java.math.BigDecimal;
-import java.time.LocalDateTime;
-import java.util.Arrays;
-import java.util.Random;
+import java.util.List;
 import java.util.UUID;
 
 public class Fixtures {
@@ -35,7 +33,7 @@ public class Fixtures {
                 MenuPrice.from(BigDecimal.valueOf(price)),
                 menuGroup(),
                 displayed,
-                MenuProducts.from(Arrays.asList(menuProductMaterial), new ProductClientImpl(productRepository))
+                MenuProducts.from(List.of(menuProductMaterial), new ProductClientImpl(productRepository))
         );
     }
 
@@ -52,7 +50,7 @@ public class Fixtures {
                 MenuPrice.from(BigDecimal.valueOf(price)),
                 menuGroup(),
                 displayed,
-                MenuProducts.from(Arrays.asList(menuProductMaterials), new ProductClientImpl(productRepository))
+                MenuProducts.from(List.of(menuProductMaterials), new ProductClientImpl(productRepository))
         );
     }
 
@@ -95,45 +93,6 @@ public class Fixtures {
                 quantity,
                 new ProductClientImpl(productRepository)
         );
-    }
-
-    public static Order order(final OrderStatus status, final String deliveryAddress) {
-        final Order order = new Order();
-        order.setId(UUID.randomUUID());
-        order.setType(OrderType.DELIVERY);
-        order.setStatus(status);
-        order.setOrderDateTime(LocalDateTime.of(2020, 1, 1, 12, 0));
-        order.setOrderLineItems(Arrays.asList(orderLineItem()));
-        order.setDeliveryAddress(deliveryAddress);
-        return order;
-    }
-
-    public static Order order(final OrderStatus status) {
-        final Order order = new Order();
-        order.setId(UUID.randomUUID());
-        order.setType(OrderType.TAKEOUT);
-        order.setStatus(status);
-        order.setOrderDateTime(LocalDateTime.of(2020, 1, 1, 12, 0));
-        order.setOrderLineItems(Arrays.asList(orderLineItem()));
-        return order;
-    }
-
-    public static Order order(final OrderStatus status, final OrderTable orderTable) {
-        final Order order = new Order();
-        order.setId(UUID.randomUUID());
-        order.setType(OrderType.EAT_IN);
-        order.setStatus(status);
-        order.setOrderDateTime(LocalDateTime.of(2020, 1, 1, 12, 0));
-        order.setOrderLineItems(Arrays.asList(orderLineItem()));
-        order.setOrderTable(orderTable);
-        return order;
-    }
-
-    public static OrderLineItem orderLineItem() {
-        final OrderLineItem orderLineItem = new OrderLineItem();
-        orderLineItem.setSeq(new Random().nextLong());
-        orderLineItem.setMenu(menu(new InMemoryProductRepository()));
-        return orderLineItem;
     }
 
     public static OrderTable orderTable() {
