+++ src/test/java/kitchenpos/delivery_orders/application/DeliveryOrderServiceTest.java
@@ -1,0 +1,247 @@
+package kitchenpos.delivery_orders.application;
+
+import kitchenpos.delivery_orders.infrastructure.FakeKitchenridersClient;
+import kitchenpos.delivery_orders.infrastructure.InMemoryDeliveryOrderRepository;
+import kitchenpos.deliveryorders.application.DeliveryOrderService;
+import kitchenpos.deliveryorders.domain.DeliveryOrder;
+import kitchenpos.deliveryorders.domain.DeliveryOrderLineItem;
+import kitchenpos.deliveryorders.domain.DeliveryOrderRepository;
+import kitchenpos.deliveryorders.domain.OrderStatus;
+import kitchenpos.menus.application.InMemoryMenuRepository;
+import kitchenpos.menus.tobe.domain.menu.MenuRepository;
+import kitchenpos.products.application.InMemoryProductRepository;
+import kitchenpos.products.tobe.domain.ProductRepository;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.EnumSource;
+import org.junit.jupiter.params.provider.MethodSource;
+import org.junit.jupiter.params.provider.NullAndEmptySource;
+
+import java.math.BigDecimal;
+import java.util.*;
+
+import static kitchenpos.Fixtures.*;
+import static kitchenpos.delivery_orders.DeliveryOrderFixtures.order;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.assertAll;
+
+class DeliveryOrderServiceTest {
+    private DeliveryOrderRepository orderRepository;
+    private MenuRepository menuRepository;
+    private ProductRepository productRepository;
+    private FakeKitchenridersClient kitchenridersClient;
+    private DeliveryOrderService orderService;
+
+    @BeforeEach
+    void setUp() {
+        orderRepository = new InMemoryDeliveryOrderRepository();
+        productRepository = new InMemoryProductRepository();
+        menuRepository = new InMemoryMenuRepository();
+        kitchenridersClient = new FakeKitchenridersClient();
+        orderService = new DeliveryOrderService(orderRepository, menuRepository, kitchenridersClient);
+    }
+
+    @DisplayName("1개 이상의 등록된 메뉴로 배달 주문을 등록할 수 있다.")
+    @Test
+    void createDeliveryOrder() {
+        final UUID menuId = menuRepository.save(menu(19_000L, true, productRepository)).getId();
+        final DeliveryOrder expected = createOrderRequest(
+                "서울시 송파구 위례성대로 2", createOrderLineItemRequest(menuId, 19_000L, 3L)
+        );
+        final DeliveryOrder actual = orderService.create(expected);
+        assertThat(actual).isNotNull();
+        assertAll(
+                () -> assertThat(actual.getId()).isNotNull(),
+                () -> assertThat(actual.getStatus()).isEqualTo(OrderStatus.WAITING),
+                () -> assertThat(actual.getOrderDateTime()).isNotNull(),
+                () -> assertThat(actual.getOrderLineItems()).hasSize(1),
+                () -> assertThat(actual.getDeliveryAddress()).isEqualTo(expected.getDeliveryAddress())
+        );
+    }
+
+    @DisplayName("메뉴가 없으면 등록할 수 없다.")
+    @MethodSource("orderLineItems")
+    @ParameterizedTest
+    void create(final List<DeliveryOrderLineItem> orderLineItems) {
+        final DeliveryOrder expected = createOrderRequest(orderLineItems);
+        assertThatThrownBy(() -> orderService.create(expected))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    private static List<Arguments> orderLineItems() {
+        return Arrays.asList(
+                null,
+                Arguments.of(Collections.emptyList()),
+                Arguments.of(Arrays.asList(createOrderLineItemRequest(INVALID_ID, 19_000L, 3L)))
+        );
+    }
+
+    @DisplayName("배달 주소가 올바르지 않으면 배달 주문을 등록할 수 없다.")
+    @NullAndEmptySource
+    @ParameterizedTest
+    void create(final String deliveryAddress) {
+        final UUID menuId = menuRepository.save(menu(19_000L, true, productRepository)).getId();
+        final DeliveryOrder expected = createOrderRequest(
+                deliveryAddress, createOrderLineItemRequest(menuId, 19_000L, 3L)
+        );
+        assertThatThrownBy(() -> orderService.create(expected))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("숨겨진 메뉴는 주문할 수 없다.")
+    @Test
+    void createNotDisplayedMenuOrder() {
+        final UUID menuId = menuRepository.save(menu(19_000L, false, productRepository)).getId();
+        final DeliveryOrder expected = createOrderRequest(
+                "서울시 송파구 위례성대로 2", createOrderLineItemRequest(menuId, 19_000L, 3L)
+        );
+        assertThatThrownBy(() -> orderService.create(expected))
+                .isInstanceOf(IllegalStateException.class);
+    }
+
+    @DisplayName("주문한 메뉴의 가격은 실제 메뉴 가격과 일치해야 한다.")
+    @Test
+    void createNotMatchedMenuPriceOrder() {
+        final UUID menuId = menuRepository.save(menu(19_000L, true, productRepository)).getId();
+        final DeliveryOrder expected = createOrderRequest(
+                "서울시 송파구 위례성대로 2", createOrderLineItemRequest(menuId, 16_000L, 3L)
+        );
+        assertThatThrownBy(() -> orderService.create(expected))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("주문을 접수한다.")
+    @Test
+    void accept() {
+        final UUID orderId = orderRepository.save(order(OrderStatus.WAITING, "서울시 송파구 위례성대로 2", productRepository)).getId();
+        final DeliveryOrder actual = orderService.accept(orderId);
+        assertThat(actual.getStatus()).isEqualTo(OrderStatus.ACCEPTED);
+    }
+
+    @DisplayName("접수 대기 중인 주문만 접수할 수 있다.")
+    @EnumSource(value = OrderStatus.class, names = "WAITING", mode = EnumSource.Mode.EXCLUDE)
+    @ParameterizedTest
+    void accept(final OrderStatus status) {
+        final UUID orderId = orderRepository.save(order(status, "서울시 송파구 위례성대로 2", productRepository)).getId();
+        assertThatThrownBy(() -> orderService.accept(orderId))
+                .isInstanceOf(IllegalStateException.class);
+    }
+
+    @DisplayName("배달 주문을 접수되면 배달 대행사를 호출한다.")
+    @Test
+    void acceptDeliveryOrder() {
+        final UUID orderId = orderRepository.save(order(OrderStatus.WAITING, "서울시 송파구 위례성대로 2", productRepository)).getId();
+        final DeliveryOrder actual = orderService.accept(orderId);
+        assertAll(
+                () -> assertThat(actual.getStatus()).isEqualTo(OrderStatus.ACCEPTED),
+                () -> assertThat(kitchenridersClient.getOrderId()).isEqualTo(orderId),
+                () -> assertThat(kitchenridersClient.getDeliveryAddress()).isEqualTo("서울시 송파구 위례성대로 2")
+        );
+    }
+
+    @DisplayName("주문을 서빙한다.")
+    @Test
+    void serve() {
+        final UUID orderId = orderRepository.save(order(OrderStatus.ACCEPTED, "서울시 송파구 위례성대로 2", productRepository)).getId();
+        final DeliveryOrder actual = orderService.serve(orderId);
+        assertThat(actual.getStatus()).isEqualTo(OrderStatus.SERVED);
+    }
+
+    @DisplayName("접수된 주문만 서빙할 수 있다.")
+    @EnumSource(value = OrderStatus.class, names = "ACCEPTED", mode = EnumSource.Mode.EXCLUDE)
+    @ParameterizedTest
+    void serve(final OrderStatus status) {
+        final UUID orderId = orderRepository.save(order(status, "서울시 송파구 위례성대로 2", productRepository)).getId();
+        assertThatThrownBy(() -> orderService.serve(orderId))
+                .isInstanceOf(IllegalStateException.class);
+    }
+
+    @DisplayName("주문을 배달한다.")
+    @Test
+    void startDelivery() {
+        final UUID orderId = orderRepository.save(order(OrderStatus.SERVED, "서울시 송파구 위례성대로 2", productRepository)).getId();
+        final DeliveryOrder actual = orderService.startDelivery(orderId);
+        assertThat(actual.getStatus()).isEqualTo(OrderStatus.DELIVERING);
+    }
+
+    @DisplayName("서빙된 주문만 배달할 수 있다.")
+    @EnumSource(value = OrderStatus.class, names = "SERVED", mode = EnumSource.Mode.EXCLUDE)
+    @ParameterizedTest
+    void startDelivery(final OrderStatus status) {
+        final UUID orderId = orderRepository.save(order(status, "서울시 송파구 위례성대로 2", productRepository)).getId();
+        assertThatThrownBy(() -> orderService.startDelivery(orderId))
+                .isInstanceOf(IllegalStateException.class);
+    }
+
+    @DisplayName("주문을 배달 완료한다.")
+    @Test
+    void completeDelivery() {
+        final UUID orderId = orderRepository.save(order(OrderStatus.DELIVERING, "서울시 송파구 위례성대로 2", productRepository)).getId();
+        final DeliveryOrder actual = orderService.completeDelivery(orderId);
+        assertThat(actual.getStatus()).isEqualTo(OrderStatus.DELIVERED);
+    }
+
+    @DisplayName("배달 중인 주문만 배달 완료할 수 있다.")
+    @EnumSource(value = OrderStatus.class, names = "DELIVERING", mode = EnumSource.Mode.EXCLUDE)
+    @ParameterizedTest
+    void completeDelivery(final OrderStatus status) {
+        final UUID orderId = orderRepository.save(order(status, "서울시 송파구 위례성대로 2", productRepository)).getId();
+        assertThatThrownBy(() -> orderService.completeDelivery(orderId))
+                .isInstanceOf(IllegalStateException.class);
+    }
+
+    @DisplayName("주문을 완료한다.")
+    @Test
+    void complete() {
+        final DeliveryOrder expected = orderRepository.save(order(OrderStatus.DELIVERED, "서울시 송파구 위례성대로 2", productRepository));
+        final DeliveryOrder actual = orderService.complete(expected.getId());
+        assertThat(actual.getStatus()).isEqualTo(OrderStatus.COMPLETED);
+    }
+
+    @DisplayName("배달 주문의 경우 배달 완료된 주문만 완료할 수 있다.")
+    @EnumSource(value = OrderStatus.class, names = "DELIVERED", mode = EnumSource.Mode.EXCLUDE)
+    @ParameterizedTest
+    void completeDeliveryOrder(final OrderStatus status) {
+        final UUID orderId = orderRepository.save(order(status, "서울시 송파구 위례성대로 2", productRepository)).getId();
+        assertThatThrownBy(() -> orderService.complete(orderId))
+                .isInstanceOf(IllegalStateException.class);
+    }
+
+    @DisplayName("주문의 목록을 조회할 수 있다.")
+    @Test
+    void findAll() {
+        orderRepository.save(order(OrderStatus.SERVED, "서울시 송파구 위례성대로 21", productRepository));
+        orderRepository.save(order(OrderStatus.DELIVERED, "서울시 송파구 위례성대로 23", productRepository));
+        final List<DeliveryOrder> actual = orderService.findAll();
+        assertThat(actual).hasSize(2);
+    }
+
+    private DeliveryOrder createOrderRequest(
+            final String deliveryAddress,
+            final DeliveryOrderLineItem... orderLineItems
+    ) {
+        final DeliveryOrder order = new DeliveryOrder();
+        order.setDeliveryAddress(deliveryAddress);
+        order.setOrderLineItems(Arrays.asList(orderLineItems));
+        return order;
+    }
+
+    private DeliveryOrder createOrderRequest(final List<DeliveryOrderLineItem> orderLineItems) {
+        final DeliveryOrder order = new DeliveryOrder();
+        order.setOrderLineItems(orderLineItems);
+        return order;
+    }
+
+    private static DeliveryOrderLineItem createOrderLineItemRequest(final UUID menuId, final long price, final long quantity) {
+        final DeliveryOrderLineItem orderLineItem = new DeliveryOrderLineItem();
+        orderLineItem.setSeq(new Random().nextLong());
+        orderLineItem.setMenuId(menuId);
+        orderLineItem.setPrice(BigDecimal.valueOf(price));
+        orderLineItem.setQuantity(quantity);
+        return orderLineItem;
+    }
+}
