+++ src/test/java/kitchenpos/eatin_orders/infrastructure/InMemoryOrderTableRepository.java
@@ -1,0 +1,26 @@
+package kitchenpos.eatin_orders.infrastructure;
+
+import kitchenpos.eatinorders.domain.ordertables.OrderTable;
+import kitchenpos.eatinorders.domain.ordertables.OrderTableRepository;
+
+import java.util.*;
+
+public class InMemoryOrderTableRepository implements OrderTableRepository {
+    private final Map<UUID, OrderTable> orderTables = new HashMap<>();
+
+    @Override
+    public OrderTable save(final OrderTable orderTable) {
+        orderTables.put(orderTable.getId(), orderTable);
+        return orderTable;
+    }
+
+    @Override
+    public Optional<OrderTable> findById(final UUID id) {
+        return Optional.ofNullable(orderTables.get(id));
+    }
+
+    @Override
+    public List<OrderTable> findAll() {
+        return new ArrayList<>(orderTables.values());
+    }
+}
