--- src/test/java/kitchenpos/eatinorders/application/InMemoryOrderRepository.java
@@ -1,35 +1,0 @@
-package kitchenpos.eatinorders.application;
-
-import kitchenpos.eatinorders.domain.Order;
-import kitchenpos.eatinorders.domain.OrderRepository;
-import kitchenpos.eatinorders.domain.OrderStatus;
-import kitchenpos.eatinorders.domain.OrderTable;
-
-import java.util.*;
-
-public class InMemoryOrderRepository implements OrderRepository {
-    private final Map<UUID, Order> orders = new HashMap<>();
-
-    @Override
-    public Order save(final Order order) {
-        orders.put(order.getId(), order);
-        return order;
-    }
-
-    @Override
-    public Optional<Order> findById(final UUID id) {
-        return Optional.ofNullable(orders.get(id));
-    }
-
-    @Override
-    public List<Order> findAll() {
-        return new ArrayList<>(orders.values());
-    }
-
-    @Override
-    public boolean existsByOrderTableAndStatusNot(final OrderTable orderTable, final OrderStatus status) {
-        return orders.values()
-                .stream()
-                .anyMatch(order -> order.getOrderTable().equals(orderTable) && order.getStatus() != status);
-    }
-}
