--- src/test/java/kitchenpos/eatinorders/application/InMemoryOrderTableRepository.java
@@ -1,26 +1,0 @@
-package kitchenpos.eatinorders.application;
-
-import kitchenpos.eatinorders.domain.OrderTable;
-import kitchenpos.eatinorders.domain.OrderTableRepository;
-
-import java.util.*;
-
-public class InMemoryOrderTableRepository implements OrderTableRepository {
-    private final Map<UUID, OrderTable> orderTables = new HashMap<>();
-
-    @Override
-    public OrderTable save(final OrderTable orderTable) {
-        orderTables.put(orderTable.getId(), orderTable);
-        return orderTable;
-    }
-
-    @Override
-    public Optional<OrderTable> findById(final UUID id) {
-        return Optional.ofNullable(orderTables.get(id));
-    }
-
-    @Override
-    public List<OrderTable> findAll() {
-        return new ArrayList<>(orderTables.values());
-    }
-}
